autoReport/requirements.txt
click==7.1.1
pandas==1.0.1
python_docx==0.8.10
